Add dead vision angle support
This allows to control a dead zone in the field of view of the boids.
It an angle expressed in degrees representing the portion in the rear of
the boid where it cannot detect the presence of other boids.

It's actually stored in the Swarm struct as it's common to all boids.
It compares the angle of the boid velocity with the position of the
neighbor. For efficiency concerns, the cosinus of the angle is
calculated and compared with what is stored into the Swarm struct. This
is much faster than calculating the angle by substracting of both angles
obtained with atan2() calls.

An angle of zero means full field of view around the boids. A value of
360 means bind boids. The default value is 60 degrees.