Add a 'type' field in Obstacle structure
This remove the 'flags' field from the Obstacle structure and replaces
it with a 'type' one of type ObstacleType which is an enum of the
possible obstacle types (IN_FIELD and WALL for now).
Also, the obstacles of type WALL are not drawn in draw_obstacles() as
they are located outside of the field.