Add a Predator in the swarm
This changes adds support for a new type of obstacle named predator.
This is a moving obstacle that follows the nearest group of Boids like
it's done for the coherence rules. Since the Predator is an obstacle,
the Boids will avoid it, giving a chasing effect.

The speed obstacle is 20% lower than the Boids speed if there is no
boids around and 20% faster if some boids can be chased.