Use DEST_OUT compositor operator for boids trail effect
Make use of CAIRO_OPERATOR_DEST_OUT for the boid trail effect as trail
traces were visible in the background with DEST_IN operator. DEST_OUT
fits better as it is used to reduce the transparency of the destination,
which is exactly what we want here.

Also add a comment explaining how the boids trail effect works.